+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_accountbillads_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_accountbillads')) {
+            return;
+        }
         Schema::table('vtiger_accountbillads', function (Blueprint $table) {
             $table->foreign(['accountaddressid'], 'fk_1_vtiger_accountbillads')->references(['accountid'])->on('vtiger_account')->onUpdate('no action')->onDelete('cascade');
         });
