+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_accountscf_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_accountscf')) {
+            return;
+        }
         Schema::table('vtiger_accountscf', function (Blueprint $table) {
             $table->foreign(['accountid'], 'fk_1_vtiger_accountscf')->references(['accountid'])->on('vtiger_account')->onUpdate('no action')->onDelete('cascade');
         });
