+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_modcommentscf_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_modcommentscf')) {
+            return;
+        }
         Schema::table('vtiger_modcommentscf', function (Blueprint $table) {
             $table->foreign(['modcommentsid'], 'fk_modcommentsid_vtiger_modcommentscf')->references(['modcommentsid'])->on('vtiger_modcomments')->onUpdate('no action')->onDelete('cascade');
         });
