+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_notescf_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_notescf')) {
+            return;
+        }
         Schema::table('vtiger_notescf', function (Blueprint $table) {
             $table->foreign(['notesid'], 'fk_notesid_vtiger_notescf')->references(['notesid'])->on('vtiger_notes')->onUpdate('no action')->onDelete('cascade');
         });
