+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_portalinfo_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_portalinfo')) {
+            return;
+        }
         Schema::table('vtiger_portalinfo', function (Blueprint $table) {
             $table->foreign(['id'], 'fk_1_vtiger_portalinfo')->references(['contactid'])->on('vtiger_contactdetails')->onUpdate('no action')->onDelete('cascade');
         });
