+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_poshipads_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_poshipads')) {
+            return;
+        }
         Schema::table('vtiger_poshipads', function (Blueprint $table) {
             $table->foreign(['poshipaddressid'], 'fk_1_vtiger_poshipads')->references(['purchaseorderid'])->on('vtiger_purchaseorder')->onUpdate('no action')->onDelete('cascade');
         });
