+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_potential_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_potential')) {
+            return;
+        }
         Schema::table('vtiger_potential', function (Blueprint $table) {
             $table->foreign(['potentialid'], 'fk_1_vtiger_potential')->references(['crmid'])->on('vtiger_crmentity')->onUpdate('no action')->onDelete('cascade');
         });
