+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_potstagehistory_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_potstagehistory')) {
+            return;
+        }
         Schema::table('vtiger_potstagehistory', function (Blueprint $table) {
             $table->foreign(['potentialid'], 'fk_1_vtiger_potstagehistory')->references(['potentialid'])->on('vtiger_potential')->onUpdate('no action')->onDelete('cascade');
         });
