+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_products_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_products')) {
+            return;
+        }
         Schema::table('vtiger_products', function (Blueprint $table) {
             $table->foreign(['productid'], 'fk_1_vtiger_products')->references(['crmid'])->on('vtiger_crmentity')->onUpdate('no action')->onDelete('cascade');
         });
