+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_project_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_project')) {
+            return;
+        }
         Schema::table('vtiger_project', function (Blueprint $table) {
             $table->foreign(['projectid'], 'fk_crmid_vtiger_project')->references(['crmid'])->on('vtiger_crmentity')->onUpdate('no action')->onDelete('cascade');
         });
