+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_quotescf_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_quotescf')) {
+            return;
+        }
         Schema::table('vtiger_quotescf', function (Blueprint $table) {
             $table->foreign(['quoteid'], 'fk_1_vtiger_quotescf')->references(['quoteid'])->on('vtiger_quotes')->onUpdate('no action')->onDelete('cascade');
         });
