+++ database/migrations/tenant/2026_01_13_135258_add_foreign_keys_to_vtiger_report_sharers_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('vtiger_report_sharers')) {
+            return;
+        }
         Schema::table('vtiger_report_sharers', function (Blueprint $table) {
             $table->foreign(['reportid'], 'vtiger_report_reportid_ibfk_4')->references(['reportid'])->on('vtiger_report')->onUpdate('no action')->onDelete('cascade');
             $table->foreign(['rsid'], 'vtiger_rolesd_rsid_ibfk_1')->references(['roleid'])->on('vtiger_role')->onUpdate('no action')->onDelete('cascade');
